Add Hello-IT scripts
Adds several scripts for use with Yoann Gini's Hello-IT application.
These scripts follow a reverse-dns naming pattern found in the default
Hello-IT scripts to avoid potential naming collisions.

Several of these scripts intended to allow arguments to be passed to
change the levels for warnings and alerts. This functionality has not
yet been implemented, in part due to issues with Hello-IT's bash script
lib, which should be fixed in the next release of Hello-IT, after which
the argument functionality will be implemented.

Hello-IT can be found here:

https://github.com/ygini/Hello-IT