Add files required to create package
Adds the build-info.plist and postinstall script required to generate a
pkg file to install scripts using munkipkg. Using a script/git hook or
to manually copy the CustomScripts folder to the proper directory and
use munkipkg to generate the pkg file.

The included postinstall script sets the correct permissions, allowing
munkipkg to be invoked without root/sudo to build the package.

The munkipkg binary can be found here:

https://github.com/munki/munki-pkg