Update pkg major version
Because of the changes to the the way Hello-IT handles scripts and the
bash script library, updated scripts no longer work with versions of
Hello-IT prior to 1.4.0. Because of these breaking changes, in the name
of an attempt at something like semantic versioning the package version
is being updated to 2.0.0. Going forward, Hello-IT Scripts will require
a Hello-IT version 1.4.0 at miniumum.